+++ node/libs/concurrency/src/limiter/mod.rs
@@ -129,7 +129,7 @@ pub struct Permit<'a> {
     ctx: &'a ctx::Ctx,
 }
 
-impl<'a> Drop for Permit<'a> {
+impl Drop for Permit<'_> {
     /// Consumes the reserved permits.
     fn drop(&mut self) {
         if self.permits == 0 {