+++ node/libs/roles/src/validator/messages/replica_commit.rs
@@ -2,6 +2,7 @@ use super::{BlockHeader, Genesis, Signed, Signers, View};
 use crate::validator;
 
 /// A commit message from a replica.
+/// WARNING: any change to this struct may invalidate preexisting signatures. See `TimeoutQC` docs.
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct ReplicaCommit {
     /// View of this message.
@@ -31,7 +32,8 @@ pub enum ReplicaCommitVerifyError {
 
 /// A Commit Quorum Certificate. It is an aggregate of signed ReplicaCommit messages.
 /// The Commit Quorum Certificate is over identical messages, so we only need one message.
-#[derive(Clone, Debug, PartialEq, Eq)]
+/// WARNING: any change to this struct may invalidate preexisting signatures. See `TimeoutQC` docs.
+#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 pub struct CommitQC {
     /// The ReplicaCommit message that the QC is for.
     pub message: ReplicaCommit,
@@ -134,18 +136,6 @@ impl CommitQC {
         self.signature
             .verify_messages(messages_and_keys)
             .map_err(CommitQCVerifyError::BadSignature)
-    }
-}
-
-impl Ord for CommitQC {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.message.view.number.cmp(&other.message.view.number)
-    }
-}
-
-impl PartialOrd for CommitQC {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
     }
 }