+++ node/libs/roles/src/validator/messages/tests/mod.rs
@@ -1,8 +1,5 @@
 use super::*;
-use crate::{
-    attester::{self, WeightedAttester},
-    validator::{self, testonly::Setup},
-};
+use crate::validator::{self, testonly::Setup};
 use rand::Rng;
 use zksync_consensus_crypto::Text;
 
@@ -18,7 +15,7 @@ mod versions;
 /// Hardcoded view.
 fn view() -> View {
     View {
-        genesis: genesis_empty_attesters().hash(),
+        genesis: genesis().hash(),
         number: ViewNumber(9136),
     }
 }
@@ -58,18 +55,6 @@ fn validator_keys() -> Vec<validator::SecretKey> {
     .collect()
 }
 
-/// Hardcoded attester secret keys.
-fn attester_keys() -> Vec<attester::SecretKey> {
-    [
-        "attester:secret:secp256k1:27cb45b1670a1ae8d376a85821d51c7f91ebc6e32788027a84758441aaf0a987",
-        "attester:secret:secp256k1:20132edc08a529e927f155e710ae7295a2a0d249f1b1f37726894d1d0d8f0d81",
-        "attester:secret:secp256k1:0946901f0a6650284726763b12de5da0f06df0016c8ec2144cf6b1903f1979a6",
-    ]
-    .iter()
-    .map(|raw| Text::new(raw).decode().unwrap())
-    .collect()
-}
-
 /// Hardcoded validator committee.
 fn validator_committee() -> Committee {
     Committee::new(
@@ -84,22 +69,8 @@ fn validator_committee() -> Committee {
     .unwrap()
 }
 
-/// Hardcoded attester committee.
-fn attester_committee() -> attester::Committee {
-    attester::Committee::new(
-        attester_keys()
-            .iter()
-            .enumerate()
-            .map(|(i, key)| WeightedAttester {
-                key: key.public(),
-                weight: i as u64 + 10,
-            }),
-    )
-    .unwrap()
-}
-
 /// Hardcoded genesis with no attesters.
-fn genesis_empty_attesters() -> Genesis {
+fn genesis() -> Genesis {
     GenesisRaw {
         chain_id: ChainId(1337),
         fork_number: ForkNumber(42),
@@ -107,22 +78,6 @@ fn genesis_empty_attesters() -> Genesis {
 
         protocol_version: ProtocolVersion(1),
         validators: validator_committee(),
-        attesters: None,
-        leader_selection: LeaderSelectionMode::Weighted,
-    }
-    .with_hash()
-}
-
-/// Hardcoded genesis with attesters.
-fn genesis_with_attesters() -> Genesis {
-    GenesisRaw {
-        chain_id: ChainId(1337),
-        fork_number: ForkNumber(42),
-        first_block: BlockNumber(2834),
-
-        protocol_version: ProtocolVersion(1),
-        validators: validator_committee(),
-        attesters: attester_committee().into(),
         leader_selection: LeaderSelectionMode::Weighted,
     }
     .with_hash()
@@ -146,7 +101,7 @@ fn replica_commit() -> ReplicaCommit {
 
 /// Hardcoded `CommitQC`.
 fn commit_qc() -> CommitQC {
-    let genesis = genesis_empty_attesters();
+    let genesis = genesis();
     let replica_commit = replica_commit();
     let mut x = CommitQC::new(replica_commit.clone(), &genesis);
     for k in validator_keys() {
@@ -160,7 +115,7 @@ fn commit_qc() -> CommitQC {
 fn replica_timeout() -> ReplicaTimeout {
     ReplicaTimeout {
         view: View {
-            genesis: genesis_empty_attesters().hash(),
+            genesis: genesis().hash(),
             number: ViewNumber(9169),
         },
         high_vote: Some(replica_commit()),
@@ -171,10 +126,10 @@ fn replica_timeout() -> ReplicaTimeout {
 /// Hardcoded `TimeoutQC`.
 fn timeout_qc() -> TimeoutQC {
     let mut x = TimeoutQC::new(View {
-        genesis: genesis_empty_attesters().hash(),
+        genesis: genesis().hash(),
         number: ViewNumber(9169),
     });
-    let genesis = genesis_empty_attesters();
+    let genesis = genesis();
     let replica_timeout = replica_timeout();
     for k in validator_keys() {
         x.add(&k.sign_msg(replica_timeout.clone()), &genesis)