+++ node/tools/src/bin/localnet_config.rs
@@ -9,6 +9,7 @@ use std::{
     os::unix::fs::PermissionsExt,
     path::PathBuf,
 };
+use zksync_concurrency::time;
 use zksync_consensus_roles::{node, validator};
 use zksync_consensus_tools::config;
 use zksync_protobuf::serde::Serialize;
@@ -75,7 +76,8 @@ fn main() -> anyhow::Result<()> {
                 .metrics_server_port
                 .map(|port| SocketAddr::new(Ipv4Addr::UNSPECIFIED.into(), port)),
             genesis: setup.genesis.clone(),
-            max_payload_size: 1000000,
+            max_payload_size: args.payload_size,
+            view_timeout: time::Duration::milliseconds(2000),
             node_key: node_keys[i].clone(),
             validator_key: validator_keys.get(i).cloned(),
             attester_key: attester_keys.get(i).cloned(),